Moving forgot password and reset password endpoints to the account namespace
