+++ src/Endatix.Api/Endpoints/Account/ForgotPassword.ForgotPasswordRequest.cs
@@ -1,4 +1,4 @@
-namespace Endatix.Api.Endpoints.Auth;
+namespace Endatix.Api.Endpoints.Account;
 
 public sealed record ForgotPasswordRequest
 {
